add models to handle criteria
- NO MIGRATION YET in case someone have a better idea of doing this.
