+++ calculator/models/admission_info.py
@@ -1,43 +1,25 @@
-"""A module to store University data."""
+"""A module to store admission criteria of each major."""
 
 from django.db import models
-
-
-# Create your models here.
-class University(models.Model):
-    """Name of different universities."""
-
-    name = models.CharField(max_length=200, null=False)
-
-    def __str__(self):
-        return self.name
-
-
-class Faculty(models.Model):
-    """Name of different faculties of a university."""
-
-    university = models.ForeignKey(University, on_delete=models.CASCADE)
-    name = models.CharField(max_length=200, null=False)
-
-    def __str__(self):
-        return f"{self.name} {self.university}"
-
-
-class Major(models.Model):
-    """Name of different majors of a faculty."""
-
-    code = models.CharField(max_length=200, null=False, unique=True)
-    name = models.CharField(max_length=200, null=False)
-    faculty = models.ForeignKey(Faculty, on_delete=models.CASCADE)
-
-    def __str__(self):
-        return f"{self.name} {self.faculty.name} {self.faculty.university.name}"
+from .university import Major
+from .exams import Exams
+from django.core.validators import MaxValueValidator
 
 
 class Criteria(models.Model):
-    """Criteria of a major."""
+    """A passing criteria and the weight of each exam."""
+
+    exam = models.ForeignKey(Exams)
+    min_score = models.FloatField()
+    weight = models.FloatField(validators=[MaxValueValidator(100)])
+
+
+class CriteriaSet(models.Model):
+    """A set of criteria for a certain major.
+
+    A major can have multiple criteria set such as this one:
+    https://course.mytcas.com/programs/10010128901101A
+    """
 
     major = models.ForeignKey(Major, on_delete=models.CASCADE)
-
-    def __str__(self):
-        return f"Criteria of {self.major.name} {self.major.faculty.name} {self.major.faculty.university.name}"
+    criteria_set = models.ManyToManyField(Criteria)
